+++ Moco/Moco/MocoGoal/MocoContactTrackingGoal.cpp
@@ -117,7 +117,6 @@ void MocoContactTrackingGoal::initializeOnModelImpl(const Model& model) const {
                         group.get_external_force_name()));
         const auto& extForce =
                 extLoads->get(group.get_external_force_name());
-        const std::string& appliedToBody = extForce.get_applied_to_body();
 
         GroupInfo groupInfo;
         for (int ic = 0; ic < group.getProperty_contact_force_paths().size();
@@ -126,39 +125,9 @@ void MocoContactTrackingGoal::initializeOnModelImpl(const Model& model) const {
             const auto& contactForce =
                     model.getComponent<SmoothSphereHalfSpaceForce>(path);
 
-            int recordOffset = findRecordOffset(group, appliedToBody);
-            const std::string& sphereBaseName =
-                    contactForce.getConnectee<PhysicalFrame>("sphere_frame")
-                            .findBaseFrame()
-                            .getName();
-            if (sphereBaseName == appliedToBody) {
-                recordOffset = 0;
-                // We want the first 3 entries in
-                // SmoothSphereHalfSpaceForce::getRecordValues(), which contain
-                // forces on the sphere.
-            }
-            if (recordOffset < 0) {
-                // The ExternalForce is not applied to the sphere's body.
-                const std::string& halfSpaceBaseName =
-                        contactForce
-                                .getConnectee<PhysicalFrame>("half_space_frame")
-                                .findBaseFrame()
-                                .getName();
-                if (halfSpaceBaseName == appliedToBody) {
-                    // We want the forces applied to the half space, which are
-                    // entries 6, 7, 8 in
-                    // SmoothSphereHalfSpaceForce::getRecordValues().
-                    recordOffset = 6;
-                }
-            }
-            OPENSIM_THROW_IF_FRMOBJ(recordOffset < 0, Exception,
-                    format("Contact force '%s' has sphere base frame '%s' "
-                           "and half space base frame '%s'; one of these "
-                           "frames should match the applied_to_body "
-                           "setting ('%s') of ExternalForce '%s'.",
-                            contactForce.getAbsolutePathString(),
-                            sphereBaseName, halfSpaceBaseName, appliedToBody,
-                            group.get_external_force_name()));
+            int recordOffset = findRecordOffset(group, contactForce,
+                    extForce.get_applied_to_body());
+
             groupInfo.contacts.push_back(
                     std::make_pair(&contactForce, recordOffset));
         }
@@ -218,41 +187,55 @@ void MocoContactTrackingGoal::initializeOnModelImpl(const Model& model) const {
 
 int MocoContactTrackingGoal::findRecordOffset(
         const MocoContactTrackingGoalGroup& group,
+        const SmoothSphereHalfSpaceForce& contactForce,
         const std::string& appliedToBody) const {
-    int recordOffset = -1;
-    const std::string& sphereBaseName =
+
+    // Is the ExternalForce applied to the sphere's body?
+    const auto& sphereBase =
             contactForce.getConnectee<PhysicalFrame>("sphere_frame")
-                    .findBaseFrame()
-                    .getName();
+                    .findBaseFrame();
+    const std::string& sphereBaseName = sphereBase.getName();
     if (sphereBaseName == appliedToBody) {
-        recordOffset = 0;
         // We want the first 3 entries in
         // SmoothSphereHalfSpaceForce::getRecordValues(), which contain
         // forces on the sphere.
+        return 0;
     }
-    if (recordOffset < 0) {
-        // The ExternalForce is not applied to the sphere's body.
-        const std::string& halfSpaceBaseName =
-                contactForce
-                        .getConnectee<PhysicalFrame>("half_space_frame")
-                        .findBaseFrame()
-                        .getName();
-        if (halfSpaceBaseName == appliedToBody) {
-            // We want the forces applied to the half space, which are
-            // entries 6, 7, 8 in
-            // SmoothSphereHalfSpaceForce::getRecordValues().
-            recordOffset = 6;
-        }
+
+    // Is the ExternalForce applied to the half space's body?
+    const auto& halfSpaceBase =
+            contactForce.getConnectee<PhysicalFrame>("half_space_frame")
+                    .findBaseFrame();
+    const std::string& halfSpaceBaseName = halfSpaceBase.getName();
+    if (halfSpaceBaseName == appliedToBody) {
+        // We want the forces applied to the half space, which are
+        // entries 6, 7, 8 in
+        // SmoothSphereHalfSpaceForce::getRecordValues().
+        return 6;
     }
-    OPENSIM_THROW_IF_FRMOBJ(recordOffset < 0, Exception,
+
+    // Check the group's alternative frames.
+    // Check each alternative frame path in the order provided. As soon as one
+    // of these paths matches the name of the sphere base frame or
+    // half space base frame, use the contact forces applied to that base frame.
+    const auto& sphereBasePath = sphereBase.getAbsolutePathString();
+    const auto& halfSpaceBasePath = halfSpaceBase.getAbsolutePathString();
+    for (int ia = 0; ia < group.getProperty_alternative_frame_paths().size();
+            ++ia) {
+        const auto& path = group.get_alternative_frame_paths(ia);
+        if (path == sphereBasePath) { return 0; }
+        if (path == halfSpaceBasePath) { return 6; }
+    }
+
+    OPENSIM_THROW_FRMOBJ(Exception,
             format("Contact force '%s' has sphere base frame '%s' "
-                   "and half space base frame '%s'; one of these "
+                  "and half space base frame '%s'. One of these "
                    "frames should match the applied_to_body "
-                   "setting ('%s') of ExternalForce '%s'.",
+                   "setting ('%s') of ExternalForce '%s', or match one of "
+                   "the alternative_frame_paths, but no match found.",
                     contactForce.getAbsolutePathString(),
                     sphereBaseName, halfSpaceBaseName, appliedToBody,
                     group.get_external_force_name()));
-
 }
 
 void MocoContactTrackingGoal::calcIntegrandImpl(